Move LoggingShell to run package
