+++ src/run/dryrun.go
@@ -1,4 +1,4 @@
-package git
+package run
 
 // DryRun implements the dry-run feature.
 // The zero value is a non-activated DryRun.
